Create merge.py v0.01
Devise an algorithm, merge, that takes as input two sorted arrays and returns a single array
with the elements of the two input arrays in sorted order. If the lengths of the input arrays
are n1 and n2, your algorithm should require O(n1 + n2) comparisons in the worst case.
Implement and test your algorithm. For example, with inputs (1, 3, 3, 6, 7) and (2, 4, 5) it
should output (1, 2, 3, 3, 4, 5, 6, 7).